requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
supabase
pytest
logging